+++ src/main/java/bfst20/mapdrawer/kdtree/KdTree.java
@@ -99,16 +99,16 @@ public class KdTree implements Serializable{
         return nearest;
     }
 
-    private double distance(Point2D point, OSMWay way) {
+    private float distance(Point2D point, OSMWay way) {
         if (way == null) {
-            return Double.MAX_VALUE; // Distance is so big anything it is compared to will be smaller
+            return Float.MAX_VALUE; // Distance is so big anything it is compared to will be smaller
         }
         
         // Keeps track of the current best distance
-        double bestDistance = Double.MAX_VALUE;
+        float bestDistance = Float.MAX_VALUE;
         
         for (OSMNode node : way.getNodes()) {
-            double distance = node.distanceSq(point);
+            float distance = node.distanceSq(point);
             
             if (bestDistance > distance) {
                 bestDistance = distance;
@@ -125,11 +125,11 @@ public class KdTree implements Serializable{
         }
 
         // Keeps track of the current best distance
-        double bestDistance = Double.MAX_VALUE;
+        float bestDistance = Float.MAX_VALUE;
         OSMNode bestNode = null;
 
         for (OSMNode node : way.getNodes()) {
-            double distance = node.distanceSq(point);
+            float distance = node.distanceSq(point);
 
             if (bestDistance > distance) {
                 bestDistance = distance;
@@ -175,10 +175,10 @@ public class KdTree implements Serializable{
             Rectangle rightBox = node.right == null ? midBox : node.right.boundingBox;
 
             // Compares all 3 boxes to each other and grabs the coordinates that will contain them all within
-            double xmin = Math.min(midBox.getXmin(), Math.min(leftBox.getXmin(), rightBox.getXmin()));
-            double xmax = Math.max(midBox.getXmax(), Math.max(leftBox.getXmax(), rightBox.getXmax()));
-            double ymin = Math.min(midBox.getYmin(), Math.min(leftBox.getYmin(), rightBox.getYmin()));
-            double ymax = Math.max(midBox.getYmax(), Math.max(leftBox.getYmax(), rightBox.getYmax()));
+            float xmin = Math.min(midBox.getXmin(), Math.min(leftBox.getXmin(), rightBox.getXmin()));
+            float xmax = Math.max(midBox.getXmax(), Math.max(leftBox.getXmax(), rightBox.getXmax()));
+            float ymin = Math.min(midBox.getYmin(), Math.min(leftBox.getYmin(), rightBox.getYmin()));
+            float ymax = Math.max(midBox.getYmax(), Math.max(leftBox.getYmax(), rightBox.getYmax()));
 
             return new Rectangle(xmin, ymin, xmax, ymax);
         }
